refactor: 홈베이스 전용 UI와 모델을 feature 레이어로 이동
