+++ src/features/homebase/ui/FourthFloor.tsx
@@ -1,15 +1,15 @@
+import type { FloorProps } from "@/features/homebase/model/floor";
 import { FloorLayout } from "./FloorLayout";
 import { Table } from "./Table";
-import { FloorProps } from "@/shared/model/homebase/Floor";
 
-export function ThirdFloor({ selectedTable, setSelectedTable }: FloorProps) {
+export function FourthFloor({ selectedTable, setSelectedTable }: FloorProps) {
   const handleSelect = (table: string) => {
     setSelectedTable(selectedTable === table ? null : table);
   };
 
   return (
-    <FloorLayout floor="3F">
-      <div className="grid grid-cols-[1fr_86px_1fr] h-[85px] lg:h-[144px] 2xl:h-[209px]">
+    <FloorLayout floor="4F">
+      <div className="grid h-[85px] grid-cols-[1fr_86px_1fr] lg:h-[144px] 2xl:h-[209px]">
         <Table
           name="테이블 1"
           capacity="6명"
@@ -18,7 +18,7 @@ export function ThirdFloor({ selectedTable, setSelectedTable }: FloorProps) {
           onClick={() => handleSelect("1")}
         />
         <div className="flex items-center justify-center border-x border-b border-sub-2">
-          <span className="text-sub-2 text-text-3 font-medium">칸막이</span>
+          <span className="text-text-3 font-medium text-sub-2">칸막이</span>
         </div>
         <Table
           name="테이블 2"
@@ -29,26 +29,20 @@ export function ThirdFloor({ selectedTable, setSelectedTable }: FloorProps) {
         />
       </div>
 
-      <div className="grid grid-cols-3 h-[85px] lg:h-[144px] 2xl:h-[209px]">
+      <div className="flex h-[85px] lg:h-[144px] 2xl:h-[209px]">
         <Table
           name="테이블 3"
           capacity="4명"
-          className="border-r border-sub-2"
+          className="flex-1 border-r border-sub-2"
           selected={selectedTable === "3"}
           onClick={() => handleSelect("3")}
         />
         <Table
-          name="테이블 5"
+          name="테이블 4"
           capacity="4명"
-          className="border-r border-sub-2"
-          selected={selectedTable === "5"}
-          onClick={() => handleSelect("5")}
-        />
-        <Table
-          name="테이블 6"
-          capacity="4명"
-          selected={selectedTable === "6"}
-          onClick={() => handleSelect("6")}
+          className="flex-1"
+          selected={selectedTable === "4"}
+          onClick={() => handleSelect("4")}
         />
       </div>
     </FloorLayout>
