+++ src/widgets/main/ui/HomebaseCard.tsx
@@ -4,9 +4,9 @@ import HomeBase from "@/shared/asset/svg/HomeBase";
 import Search from "@/shared/asset/svg/Search";
 import TextField from "@/shared/ui/textField";
 import { TextButton } from "@/shared/ui/Button/TextButton";
-import { SecondFloor } from "@/shared/ui/homebase/SecondFloor";
-import { ThirdFloor } from "@/shared/ui/homebase/ThirdFloor";
-import { FourthFloor } from "@/shared/ui/homebase/FourthFloor";
+import { SecondFloor } from "@/features/homebase/ui/SecondFloor";
+import { ThirdFloor } from "@/features/homebase/ui/ThirdFloor";
+import { FourthFloor } from "@/features/homebase/ui/FourthFloor";
 import StudentSearch from "@/features/homebase/ui/StudentSearch";
 import SelectedStudent from "@/features/homebase/ui/SelectedStudent";
 import { ReservationTableItem } from "@/entities/school/ui/ReservationTableItem";
